:sparkles: Feat: notice, suggestion 통합
통합함
