+++ src/components/common/이제 필요없는.../commonList/CommonList.jsx
@@ -1,17 +1,16 @@
 import React, { useState, useEffect } from "react";
-import { useLocation, useNavigate, useParams } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import * as S from "./style";
 
 // 컴포넌트
-import Selector from "../selector/Selector";
-import Paging from "../paging/Paging";
+import Selector from "../../selector/Selector";
+import Paging from "../../paging/Paging";
 import { useRecoilState } from "recoil";
-import { userState } from "../../../context/authState";
-import NoPage from "../noPage/NoPage";
+import { userState } from "../../../../context/authState";
+import NoPage from "../../noPage/NoPage";
 
-const List = ({
-  use,
+const CommonList = ({
   data,
   url,
   writeUrl,
@@ -22,8 +21,6 @@ const List = ({
   setCurrentPage,
   count
 }) => {
-  const thList = ["번호", "제목", "서비스명", "등록일시", "좋아요", "조회수"];
-  // 회원 정보
   const [userInfo, setUserInfo] = useRecoilState(userState);
 
   const navigate = useNavigate();
@@ -36,31 +33,6 @@ const List = ({
   const handlePageChange = pageNumber => {
     setCurrentPage(pageNumber);
   };
-  const location = useLocation();
-  const decodeName = decodeURI(location.pathname.split("/")[2]);
-  console.log(decodeName);
-
-  const [isMobile, setisMobile] = useState(false);
-
-  //윈도우가 550px 이하면  모바일버전을 연다
-  const resizingHandler = () => {
-    if (window.innerWidth < 550) {
-      setisMobile(true);
-    } else {
-      setisMobile(false);
-    }
-  };
-
-  useEffect(() => {
-    if (window.innerWidth <= 550) {
-      setisMobile(true);
-    }
-    window.addEventListener("resize", resizingHandler);
-
-    return () => {
-      window.removeEventListener("resize", resizingHandler);
-    };
-  });
 
   return (
     <>
@@ -73,7 +45,7 @@ const List = ({
                 !userInfo
                   ? navigate("/login")
                   : navigate(writeUrl, {
-                      state: { category: "tip", ai: decodeName }
+                      state: { category: "common", ai: "" }
                     });
               }}
             >
@@ -81,7 +53,6 @@ const List = ({
               글쓰기
             </S.AiServiceDetailTipHeaderWriteContent>
           </S.AiServiceDetailTipHeaderWrite>
-
           <S.AiServiceDetailTipHeaderSort>
             <Selector
               options={SelectorOption}
@@ -89,24 +60,16 @@ const List = ({
             />
           </S.AiServiceDetailTipHeaderSort>
         </S.AiServiceDetailTipHeader>
-
         <S.AiServiceDetailTipLine></S.AiServiceDetailTipLine>
         {/* 데이터 목록 */}
         {data && data.length > 0 ? (
           <S.AiServiceDetailTipTable>
             <S.AiServiceDetailTipTableThead>
               <S.AiServiceDetailTipTableTr>
-                {thList.map((thTitle, idx) => (
-                  <S.AiServiceDetailTipTableTh key={idx}>
-                    {번호}
-                  </S.AiServiceDetailTipTableTh>
-                ))}
                 <S.AiServiceDetailTipTableTh>번호</S.AiServiceDetailTipTableTh>
-
                 <S.AiServiceDetailTipTableTh>제목</S.AiServiceDetailTipTableTh>
-                <S.AiServiceDetailTipTableTh>
-                  서비스명
-                </S.AiServiceDetailTipTableTh>
+                <S.AiServiceDetailTipTableTh></S.AiServiceDetailTipTableTh>
+
                 <S.AiServiceDetailTipTableTh>
                   등록일시
                 </S.AiServiceDetailTipTableTh>
@@ -124,25 +87,19 @@ const List = ({
                   key={data.id}
                   onClick={() => navigate(`${url}${data.id}`)}
                 >
-                  {isMobile ? (
-                    <></>
-                  ) : (
-                    <S.AiServiceDetailTipTableTd>
-                      {currentOption === "popular" || currentOption === "like"
-                        ? idx + 1 + (currentPage - 1) * itemsPerPage
-                        : count - idx - (currentPage - 1) * itemsPerPage}
-                    </S.AiServiceDetailTipTableTd>
-                  )}
-
+                  <S.AiServiceDetailTipTableTd>
+                    {currentOption === "popular" || currentOption === "like"
+                      ? idx + 1 + (currentPage - 1) * itemsPerPage
+                      : count - idx - (currentPage - 1) * itemsPerPage}
+                  </S.AiServiceDetailTipTableTd>
                   <S.AiServiceDetailTipTableTdTitle>
                     {data.title}{" "}
                     <strong style={{ fontSize: "1.6rem", color: "#4285F4" }}>
                       [{data.comments_cnt}]
                     </strong>
                   </S.AiServiceDetailTipTableTdTitle>
-                  <S.AiServiceDetailTipTableTd>
-                    {data.ai}
-                  </S.AiServiceDetailTipTableTd>
+
+                  <S.AiServiceDetailTipTableTd></S.AiServiceDetailTipTableTd>
                   <S.AiServiceDetailTipTableTd>
                     {data.created_at.split(" ")[0]}
                   </S.AiServiceDetailTipTableTd>
@@ -175,4 +132,4 @@ const List = ({
   );
 };
 
-export default List;
+export default CommonList;
