+++ src/components/common/이제 필요없는.../suggestionList/SuggestionList.jsx
