+++ src/components/common/이제 필요없는.../tipsList/TipsList.jsx
